% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{scanse}
\alias{scanse}
\title{Identify super enhancers}
\usage{
scanse(d, distance = 125000, min_peak_nums = 3)
}
\arguments{
\item{d}{A dataframe from output of chipseeker}

\item{distance}{The distance for merging two adjacent peaks}

\item{min_peak_nums}{min peak numbers}
}
\value{
A dataframe
}
\description{
Identify super enhancers
}
\examples{
scanse(d)
}
